backend/requirements.txt
Django==4.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-environ==0.12.0
psycopg2==2.9.10
PyJWT==2.10.1
requests==2.32.3
cryptography==44.0.2